Cleaned up old string extension methods
Collapsed two similar files into one, removed a bunch of old (and faulty) "CSV" parsing methods
